+++ pkg/proxy/clients.go
@@ -1,28 +1,56 @@
 package proxy
 
 import (
-	"github.com/emvi/logbuch"
+	"log/slog"
+
 	pirsch "github.com/pirsch-analytics/pirsch-go-sdk/v2/pkg"
 )
 
 var (
-	clients []*pirsch.Client
+	clients []client
 )
+
+type client struct {
+	api    *pirsch.Client
+	filter []FilterFunc
+}
 
 // SetupClients initializes all configured clients.
 func SetupClients() {
 	for _, c := range config.Clients {
-		logbuch.Info("Adding client", logbuch.Fields{"id": c.ID, "base_url": config.BaseURL})
-		client := pirsch.NewClient(c.ID, c.Secret, &pirsch.ClientConfig{
+		slog.Info("Adding client", "id", c.ID, "base_url", config.BaseURL)
+		pirschClient := pirsch.NewClient(c.ID, c.Secret, &pirsch.ClientConfig{
 			BaseURL: config.BaseURL,
 		})
 
 		if c.ID != "" {
-			if _, err := client.Domain(); err != nil {
-				logbuch.Fatal("Error connecting client", logbuch.Fields{"err": err})
+			if _, err := pirschClient.Domain(); err != nil {
+				slog.Error("Error connecting client", "err", err)
+				panic(err)
 			}
 		}
 
-		clients = append(clients, client)
+		clients = append(clients, client{
+			api:    pirschClient,
+			filter: createFilter(c.Filter),
+		})
 	}
+}
+
+func createFilter(config ClientFilter) []FilterFunc {
+	f := make([]FilterFunc, 0)
+
+	if len(config.Hostname) > 0 {
+		f = append(f, NewHostnameFilter(config.Hostname))
+	}
+
+	if len(config.Path) > 0 {
+		f = append(f, NewPathFilter(config.Path))
+	}
+
+	if len(config.IdentificationCode) > 0 {
+		f = append(f, NewIdentificationCodeFilter(config.IdentificationCode))
+	}
+
+	return f
 }